+++ src/test_helpers/mod.rs
@@ -1,4 +1,4 @@
-use async_std::prelude::*;
+use tokio::prelude::*;
 use chrono::Utc;
 use serde::de::DeserializeOwned;
 use serde_json::json;